+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT4A32HR7G0AVX26GK1MTCB4"
+        "01KT4CMJ5PVPGPR3RXFMX8VV9W"
     )
     r"""The unique identifier of the data source"""