+++ packages/orq-rc/src/orq_ai_sdk/models/createeval_guardrail_config_evals_response_200_applicationjson_responsebody_4_type.py
@@ -1155,9 +1155,9 @@ class Typescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-02T13:58:48.736Z"
+    created: Optional[str] = "2026-06-02T14:43:10.201Z"
 
-    updated: Optional[str] = "2026-06-02T13:58:48.736Z"
+    updated: Optional[str] = "2026-06-02T14:43:10.201Z"
 
     guardrail_config: OptionalNullable[
         CreateEvalResponseBodyEvalsResponse200ApplicationJSONGuardrailConfig
@@ -1398,9 +1398,9 @@ class Ragas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-06-02T13:58:48.736Z"
+    created: Optional[str] = "2026-06-02T14:43:10.201Z"
 
-    updated: Optional[str] = "2026-06-02T13:58:48.736Z"
+    updated: Optional[str] = "2026-06-02T14:43:10.201Z"
 
     guardrail_config: OptionalNullable[
         CreateEvalResponseBodyEvalsResponse200GuardrailConfig
@@ -2130,9 +2130,9 @@ class CreateEvalResponseBodyFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-02T13:58:48.736Z"
+    created: Optional[str] = "2026-06-02T14:43:10.201Z"
 
-    updated: Optional[str] = "2026-06-02T13:58:48.736Z"
+    updated: Optional[str] = "2026-06-02T14:43:10.201Z"
 
     guardrail_config: OptionalNullable[
         CreateEvalResponseBodyEvalsResponseGuardrailConfig