+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT4A32HVWGTZ598RB0JZP27A"
+        "01KT4CMJ5S0ND1V21XWTCVCH1Q"
     )
     r"""The unique identifier of the data source"""