+++ packages/orq-rc/RELEASES.md
@@ -8399,3 +8399,13 @@ Based on:
 - [python v4.10.0-rc.72] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.72] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.72 - packages/orq-rc
+
+## 2026-06-02 14:46:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.73] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.73] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.73 - packages/orq-rc